Fix : stack allocation - Add : dump instruction

+++ srcs/Factory.cpp
@@ -38,7 +38,7 @@ IOperand const *Factory::createInt8(std::string const &value) const
 
 IOperand const *Factory::createInt16(std::string const &value) const
 {
-	return new Operand<int8_t>(Int16, value);
+	return new Operand<int16_t>(Int16, value);
 }
 
 IOperand const *Factory::createInt32(std::string const &value) const
